Create Jobs in TaaS using user token instead of M2M
