+++ src/events/projects/index.js
@@ -5,7 +5,6 @@ import _ from 'lodash';
 import Joi from 'joi';
 import Promise from 'bluebird';
 import config from 'config';
-import axios from 'axios';
 import util from '../../util';
 import models from '../../models';
 import { createPhaseTopic } from '../projectPhases';
@@ -14,27 +13,6 @@ import { PROJECT_STATUS, REGEX, TIMELINE_REFERENCES } from '../../constants';
 const ES_PROJECT_INDEX = config.get('elasticsearchConfig.indexName');
 const ES_PROJECT_TYPE = config.get('elasticsearchConfig.docType');
 const eClient = util.getElasticSearchClient();
-
-/**
-  * creates taas job
-  * @param {Object} data the job data
-  * @return {Object} the job created
-  */
-const createTaasJob = async (data) => {
-  const token = await util.getM2MToken();
-  const headers = {
-    'Content-Type': 'application/json',
-    Authorization: `Bearer ${token}`,
-  };
-  const res = await axios
-    .post(config.taasJobApiUrl, data, { headers })
-    .catch((err) => {
-      const error = new Error();
-      error.message = _.get(err, 'response.data.message', error.message);
-      throw error;
-    });
-  return res.data;
-};
 
 /**
  * Payload for deprecated BUS events like `connect.notification.project.updated`.
@@ -185,42 +163,6 @@ async function projectCreatedKafkaHandler(app, topic, payload) {
     );
     await Promise.all(topicPromises);
     app.logger.debug('Topics for phases are successfully created.');
-  }
-  try {
-    if (project.type === 'talent-as-a-service') {
-      const jobs = _.get(project, 'details.taasDefinition.taasJobs');
-      if (!jobs || !jobs.length) {
-        app.logger.debug(`no jobs found in the project id: ${project.id}`);
-        return;
-      }
-      app.logger.debug(`${jobs.length} jobs found in the project id: ${project.id}`);
-      await Promise.all(
-        _.map(
-          jobs,
-          (job) => {
-            // make sure that skills would be unique in the list and only include ones with 'skillId' (actually they all suppose to be with skillId)
-            const skills = _.chain(job.skills).map('skillId').uniq().compact()
-              .value();
-            return createTaasJob({
-              projectId: project.id,
-              title: job.title,
-              description: job.description,
-              skills,
-              numPositions: Number(job.people),
-              resourceType: _.get(job, 'role.value', ''),
-              rateType: 'weekly', // hardcode for now
-              workload: _.get(job, 'workLoad.title', '').toLowerCase(),
-            }).then((createdJob) => {
-              app.logger.debug(`jobId: ${createdJob.id} job created with title "${createdJob.title}"`);
-            }).catch((err) => {
-              app.logger.error(`Unable to create job with title "${job.title}": ${err.message}`);
-            });
-          },
-        ),
-      );
-    }
-  } catch (error) {
-    app.logger.error(`Error while creating TaaS jobs: ${error}`);
   }
 }
